+++ Specs/Core/BoundingSphereSpec.js
@@ -913,5 +913,36 @@ defineSuite([
         expect(computedVolume).toEqualEpsilon(expectedVolume, CesiumMath.EPSILON6);
     });
 
+    it('clampToBounds given input value outside of sphere', function() {
+        var sphere = new BoundingSphere(Cartesian3.ZERO, 1.0);
+        var testLocation = new Cartesian3(2, 2, 2);
+        var result = sphere.clampToBounds(testLocation, new Cartesian3());
+
+        expect(result).not.toEqual(testLocation);
+
+        var result2 = sphere.clampToBounds(testLocation, new Cartesian3());
+        expect(result2).toEqual(result);
+
+        var finalCheck = new Cartesian3(2 / Math.sqrt(12), 2 / Math.sqrt(12), 2 / Math.sqrt(12));
+        expect(result).toEqual(finalCheck);
+    });
+
+    it('clampToBounds given input value is center of sphere', function() {
+        var sphere = new BoundingSphere(Cartesian3.ZERO, 3.0);
+        var testLocation = new Cartesian3(0, 0, 0);
+        var result = sphere.clampToBounds(testLocation, new Cartesian3());
+        sphere.clampToBounds(testLocation, result);
+
+        expect(result).toEqual(testLocation);
+    });
+
+    it('clampToBounds given input value inside of sphere but not center', function() {
+        var sphere = new BoundingSphere(Cartesian3.ZERO, 3.0);
+        var testLocation = new Cartesian3(1, 1, 1);
+        var result = sphere.clampToBounds(testLocation, new Cartesian3());
+
+        expect(result).toEqual(testLocation);
+    });
+
     createPackableSpecs(BoundingSphere, new BoundingSphere(new Cartesian3(1.0, 2.0, 3.0), 4.0), [1.0, 2.0, 3.0, 4.0]);
 });